aura-your-ai-executive-assistant-main/aura-agent/requirements.txt
# AURA Agent Backend Requirements

# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Providers
openai>=1.0.0
ollama>=0.1.0

# Notion Integration
notion-client>=2.0.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Voice Input
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Optional: For enhanced functionality
# pydantic>=2.0.0  # Data validation
# google-auth==2.27.0
# google-auth-oauthlib==1.2.0
# google-auth-httplib2==0.2.0
# google-api-python-client==2.116.0

# Development
pytest==7.4.3
black==23.12.1
